refactor: moving the id parsing functions out to their own package
Updating this to follow the final pattern used by other packages
