+++ azurerm/internal/services/network/parse/route_filter_test.go
@@ -1,14 +1,14 @@
-package network
+package parse
 
 import (
 	"testing"
 )
 
-func TestParseVPNGateway(t *testing.T) {
+func TestParseRouteFilter(t *testing.T) {
 	testData := []struct {
 		Name     string
 		Input    string
-		Expected *VPNGatewayResourceID
+		Expected *RouteFilterId
 	}{
 		{
 			Name:     "Empty",
@@ -16,19 +16,19 @@ func TestParseVPNGateway(t *testing.T) {
 			Expected: nil,
 		},
 		{
-			Name:     "No VPN Gateways Segment",
+			Name:     "No Route Filter Segment",
 			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
 			Expected: nil,
 		},
 		{
-			Name:     "No VPN Gateways Value",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/vpnGateways/",
+			Name:     "No Route Filter Value",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/routeFilters/",
 			Expected: nil,
 		},
 		{
 			Name:  "Completed",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/vpnGateways/example",
-			Expected: &VPNGatewayResourceID{
+			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/routeFilters/example",
+			Expected: &RouteFilterId{
 				Name:          "example",
 				ResourceGroup: "foo",
 			},
@@ -38,7 +38,7 @@ func TestParseVPNGateway(t *testing.T) {
 	for _, v := range testData {
 		t.Logf("[DEBUG] Testing %q", v.Name)
 
-		actual, err := ParseVPNGatewayID(v.Input)
+		actual, err := RouteFilterID(v.Input)
 		if err != nil {
 			if v.Expected == nil {
 				continue
