+++ azurerm/internal/services/network/parse/route_table.go
@@ -1,4 +1,4 @@
-package network
+package parse
 
 import (
 	"fmt"
@@ -6,21 +6,18 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 )
 
-// NOTE: there's some nice things we can do with this around validation
-// since these top level objects exist
-
-type RouteTableResourceID struct {
+type RouteTableId struct {
 	ResourceGroup string
 	Name          string
 }
 
-func ParseRouteTableID(input string) (*RouteTableResourceID, error) {
+func RouteTableID(input string) (*RouteTableId, error) {
 	id, err := azure.ParseAzureResourceID(input)
 	if err != nil {
 		return nil, fmt.Errorf("[ERROR] Unable to parse Route Table ID %q: %+v", input, err)
 	}
 
-	routeTable := RouteTableResourceID{
+	routeTable := RouteTableId{
 		ResourceGroup: id.ResourceGroup,
 	}
