+++ azurerm/internal/services/network/parse/route_table_test.go
@@ -1,0 +1,56 @@
+package parse
+
+import "testing"
+
+func TestParseRouteTable(t *testing.T) {
+	testData := []struct {
+		Name     string
+		Input    string
+		Expected *RouteTableId
+	}{
+		{
+			Name:     "Empty",
+			Input:    "",
+			Expected: nil,
+		},
+		{
+			Name:     "No Route Table Segment",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
+			Expected: nil,
+		},
+		{
+			Name:     "No Route Table Value",
+			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/routeTables/",
+			Expected: nil,
+		},
+		{
+			Name:  "Completed",
+			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/routeTables/example",
+			Expected: &RouteTableId{
+				Name:          "example",
+				ResourceGroup: "foo",
+			},
+		},
+	}
+
+	for _, v := range testData {
+		t.Logf("[DEBUG] Testing %q", v.Name)
+
+		actual, err := RouteTableID(v.Input)
+		if err != nil {
+			if v.Expected == nil {
+				continue
+			}
+
+			t.Fatalf("Expected a value but got an error: %s", err)
+		}
+
+		if actual.Name != v.Expected.Name {
+			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
+		}
+
+		if actual.ResourceGroup != v.Expected.ResourceGroup {
+			t.Fatalf("Expected %q but got %q for ResourceGroup", v.Expected.ResourceGroup, actual.ResourceGroup)
+		}
+	}
+}
