+++ azurerm/internal/services/network/parse/virtual_wan.go
@@ -1,4 +1,4 @@
-package network
+package parse
 
 import (
 	"fmt"
@@ -6,18 +6,18 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 )
 
-type VirtualWanResourceID struct {
+type VirtualWanId struct {
 	ResourceGroup string
 	Name          string
 }
 
-func ParseVirtualWanID(input string) (*VirtualWanResourceID, error) {
+func VirtualWanID(input string) (*VirtualWanId, error) {
 	id, err := azure.ParseAzureResourceID(input)
 	if err != nil {
 		return nil, fmt.Errorf("[ERROR] Unable to parse Virtual Wan ID %q: %+v", input, err)
 	}
 
-	virtualWan := VirtualWanResourceID{
+	virtualWan := VirtualWanId{
 		ResourceGroup: id.ResourceGroup,
 	}
 
@@ -30,17 +30,4 @@ func ParseVirtualWanID(input string) (*VirtualWanResourceID, error) {
 	}
 
 	return &virtualWan, nil
-}
-
-func ValidateVirtualWanID(i interface{}, k string) (warnings []string, errors []error) {
-	v, ok := i.(string)
-	if !ok {
-		return nil, []error{fmt.Errorf("expected type of %q to be string", k)}
-	}
-
-	if _, err := ParseVirtualWanID(v); err != nil {
-		return nil, []error{err}
-	}
-
-	return nil, nil
 }
