+++ azurerm/internal/services/network/parse/virtual_wan_test.go
@@ -1,14 +1,12 @@
-package network
+package parse
 
-import (
-	"testing"
-)
+import "testing"
 
 func TestParseVirtualWan(t *testing.T) {
 	testData := []struct {
 		Name     string
 		Input    string
-		Expected *VirtualWanResourceID
+		Expected *VirtualWanId
 	}{
 		{
 			Name:     "Empty",
@@ -28,7 +26,7 @@ func TestParseVirtualWan(t *testing.T) {
 		{
 			Name:  "Completed",
 			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualWans/example",
-			Expected: &VirtualWanResourceID{
+			Expected: &VirtualWanId{
 				Name:          "example",
 				ResourceGroup: "foo",
 			},
@@ -38,7 +36,7 @@ func TestParseVirtualWan(t *testing.T) {
 	for _, v := range testData {
 		t.Logf("[DEBUG] Testing %q", v.Name)
 
-		actual, err := ParseVirtualWanID(v.Input)
+		actual, err := VirtualWanID(v.Input)
 		if err != nil {
 			if v.Expected == nil {
 				continue
@@ -53,45 +51,6 @@ func TestParseVirtualWan(t *testing.T) {
 
 		if actual.ResourceGroup != v.Expected.ResourceGroup {
 			t.Fatalf("Expected %q but got %q for ResourceGroup", v.Expected.ResourceGroup, actual.ResourceGroup)
-		}
-	}
-}
-
-func TestValidateVirtualWan(t *testing.T) {
-	testData := []struct {
-		Name  string
-		Input string
-		Valid bool
-	}{
-		{
-			Name:  "Empty",
-			Input: "",
-			Valid: false,
-		},
-		{
-			Name:  "No Virtual Wans Segment",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
-			Valid: false,
-		},
-		{
-			Name:  "No Virtual Wans Value",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualWans/",
-			Valid: false,
-		},
-		{
-			Name:  "Completed",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualWans/example",
-			Valid: true,
-		},
-	}
-
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Input)
-
-		_, errors := ValidateVirtualWanID(v.Input, "virtual_wan_id")
-		isValid := len(errors) == 0
-		if v.Valid != isValid {
-			t.Fatalf("Expected %t but got %t", v.Valid, isValid)
 		}
 	}
 }
