+++ azurerm/internal/services/network/validate/virtual_hub.go
@@ -1,0 +1,20 @@
+package validate
+
+import (
+	"fmt"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/network/parse"
+)
+
+func VirtualHubID(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		return nil, []error{fmt.Errorf("expected type of %q to be string", k)}
+	}
+
+	if _, err := parse.VirtualHubID(v); err != nil {
+		return nil, []error{err}
+	}
+
+	return nil, nil
+}
