--- azurerm/internal/services/network/virtual_hub_test.go
@@ -1,97 +1,0 @@
-package network
-
-import (
-	"testing"
-)
-
-func TestParseVirtualHub(t *testing.T) {
-	testData := []struct {
-		Name     string
-		Input    string
-		Expected *VirtualHubResourceID
-	}{
-		{
-			Name:     "Empty",
-			Input:    "",
-			Expected: nil,
-		},
-		{
-			Name:     "No Virtual Hubs Segment",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
-			Expected: nil,
-		},
-		{
-			Name:     "No Virtual Hubs Value",
-			Input:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualHubs/",
-			Expected: nil,
-		},
-		{
-			Name:  "Completed",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualHubs/example",
-			Expected: &VirtualHubResourceID{
-				Name:          "example",
-				ResourceGroup: "foo",
-			},
-		},
-	}
-
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Name)
-
-		actual, err := ParseVirtualHubID(v.Input)
-		if err != nil {
-			if v.Expected == nil {
-				continue
-			}
-
-			t.Fatalf("Expected a value but got an error: %s", err)
-		}
-
-		if actual.Name != v.Expected.Name {
-			t.Fatalf("Expected %q but got %q for Name", v.Expected.Name, actual.Name)
-		}
-
-		if actual.ResourceGroup != v.Expected.ResourceGroup {
-			t.Fatalf("Expected %q but got %q for ResourceGroup", v.Expected.ResourceGroup, actual.ResourceGroup)
-		}
-	}
-}
-
-func TestValidateVirtualHub(t *testing.T) {
-	testData := []struct {
-		Name  string
-		Input string
-		Valid bool
-	}{
-		{
-			Name:  "Empty",
-			Input: "",
-			Valid: false,
-		},
-		{
-			Name:  "No Virtual Hubs Segment",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
-			Valid: false,
-		},
-		{
-			Name:  "No Virtual Hubs Value",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualHubs/",
-			Valid: false,
-		},
-		{
-			Name:  "Completed",
-			Input: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/virtualHubs/example",
-			Valid: true,
-		},
-	}
-
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Input)
-
-		_, errors := ValidateVirtualHubID(v.Input, "virtual_hub_id")
-		isValid := len(errors) == 0
-		if v.Valid != isValid {
-			t.Fatalf("Expected %t but got %t", v.Valid, isValid)
-		}
-	}
-}
